requirements.txt
requests
pyperclip
python-dotenv
selenium
bs4
tldextract